Move sparse matrix and tower in Strong_collapse folder
