+++ src/Strong_collapse/test/test_flag_complex_sparse_matrix.cpp
@@ -27,7 +27,7 @@
 #include <vector>
 #include <tuple>
 
-#include <gudhi/Flag_complex_sparse_matrix.h>
+#include <gudhi/Strong_collapse/Flag_complex_sparse_matrix.h>
 #include <gudhi/Unitary_tests_utils.h>
 
 // Type definitions
